+++ app/TypiCMS/Modules/Pages/Controllers/Admin/PagesController.php
@@ -107,11 +107,11 @@ class PagesController extends BaseController {
 		$data = Input::all();
 
 		// add checkboxes data
-		$data['rss_enabled']      = Input::get('rss_enabled');
-		$data['comments_enabled'] = Input::get('comments_enabled');
-		$data['is_home']          = Input::get('is_home');
+		$data['rss_enabled']      = (int) Input::get('rss_enabled');
+		$data['comments_enabled'] = (int) Input::get('comments_enabled');
+		$data['is_home']          = (int) Input::get('is_home');
 		foreach (Config::get('app.locales') as $locale) {
-			$data[$locale]['status'] = Input::get($locale.'.status');
+			$data[$locale]['status'] = (int) Input::get($locale.'.status');
 		}
 
 		if ( $this->form->update( $data ) ) {
